+++ src/states/mod.rs
@@ -1,0 +1,4 @@
+// Namespace states
+// Contains modules related to drawable states
+
+include!("./state_machine.rs");
